+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Removed
+
+* Removed resource `version.properties`. Version is set at generate-sources phase.
+* Removed class `com.unzer.payment.exceptions.PropertiesException`
+* Removed deprecated method `com.unzer.payment.util.SDKInfo::getVersion()`. Please, use `com.unzer.payment.util.SDKInfo.VERSION` instead 
+
 ## [3.1.1][3.1.1]
 
 This release updates dependencies with security issues